move all tests to test package
