+++ dynomitemanager/src/test/java/com/netflix/dynomitemanager/sidecore/utils/test/FloridaHealthCheckHandlerTest.java
@@ -1,8 +1,9 @@
-package com.netflix.dynomitemanager.sidecore.utils;
+package com.netflix.dynomitemanager.sidecore.utils.test;
 
 import org.junit.Assert;
 import org.junit.Test;
 import com.netflix.dynomitemanager.InstanceState;
+import com.netflix.dynomitemanager.sidecore.utils.FloridaHealthCheckHandler;
 
 /**
  * FloridaHealthCheckHandler unit tests
