+++ dynomitemanager/src/test/java/com/netflix/dynomitemanager/sidecore/utils/test/RedisInfoParserTest.java
@@ -1,4 +1,4 @@
-package com.netflix.dynomitemanager.sidecore.utils;
+package com.netflix.dynomitemanager.sidecore.utils.test;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -8,6 +8,8 @@ import java.util.Map;
 
 import org.junit.Assert;
 import org.junit.Test;
+
+import com.netflix.dynomitemanager.sidecore.utils.RedisInfoParser;
 
 /**
  *
